refactor: MapViewModel을 MapViewModel과 SearchViewModel로 분리
